requirements.txt
alembic==1.17.2
bcrypt==5.0.0
blinker==1.9.0
click==8.3.1
Deprecated==1.3.1
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-Limiter==4.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
greenlet==3.2.4
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.6.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mysqlclient==2.2.7
ordered-set==4.1.0
packaging==25.0
Pygments==2.19.2
python-dotenv==1.2.1
rich==14.2.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
wrapt==2.0.1
WTForms==3.2.1
